modification

+++ server3.py
@@ -121,7 +121,7 @@ def sendframe(framedata):
    
     numpy.copyto(dbuf[:480*(3+rgba)], numpy.frombuffer(framedata, dtype=numpy.uint8))
 
-    display.display(dbuf.ctypes.data_as(POINTER(c_uint8)), ctypes.c_uint8(BRIGHTNESS), ctypes.c_uint8(rgba))
+    display.display(dbuf.ctypes.data_as(POINTER(c_uint8)), ctypes.c_float(BRIGHTNESS), ctypes.c_uint8(rgba))
 
 class UDPServer:
